4_frameworks/django_daphne_websocket/django_app/serializers.py
from rest_framework import serializers
from django_app import models


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Notification
        fields = '__all__'
