4_frameworks/django_daphne_websocket/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.4
django-redis==5.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
packaging==23.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2023.3
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
zope.interface==5.5.2
